split of implementation from constructor
