+++ src/Of.php
@@ -6,35 +6,12 @@ namespace Innmind\Validation;
 use Innmind\Immutable\Validation;
 
 /**
- * @template-covariant A
- * @template-covariant B
- * @implements Constraint\Implementation<A, B>
- * @implements Constraint\Provider<A, B>
  * @psalm-immutable
  */
-final class Of implements Constraint\Implementation, Constraint\Provider
+final class Of
 {
-    /** @var pure-callable(A): Validation<Failure, B> */
-    private $assert;
-
-    /**
-     * @param pure-callable(A): Validation<Failure, B> $assert
-     */
-    private function __construct(callable $assert)
+    private function __construct()
     {
-        $this->assert = $assert;
-    }
-
-    #[\Override]
-    public function __invoke(mixed $value): Validation
-    {
-        return ($this->assert)($value);
-    }
-
-    #[\Override]
-    public function toConstraint(): Constraint
-    {
-        return Constraint::build($this);
     }
 
     /**
@@ -44,10 +21,10 @@ final class Of implements Constraint\Implementation, Constraint\Provider
      *
      * @param pure-callable(T): Validation<Failure, U> $assert
      *
-     * @return self<T, U>
+     * @return Constraint<T, U>
      */
-    public static function callable(callable $assert): self
+    public static function callable(callable $assert): Constraint
     {
-        return new self($assert);
+        return Constraint::of($assert);
     }
 }
